修改打印功能，之前打印格式用的usize，改成了isize 完成了第一部分作业

+++ src/process/proc/mod.rs
@@ -542,7 +542,7 @@ impl Proc {
 
         if a7 < SYSCALL_NAMES.len() && (self.trace_mask >> a7)&1 != 0 {
             let pid = self.excl.lock().pid;
-            println!("{}: syscall {} -> {}", pid, SYSCALL_NAMES[a7], ret);
+            println!("{}: syscall {} -> {}", pid, SYSCALL_NAMES[a7], ret as isize);
         }
 
     }
